requirements.txt
streamlit
deepface
PIL
cv2
matplotlib.pyplot
matplotlib.patches
sklearn
requests
numpy